[mpim] Single precision configure flag
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/681

This completes the milestone for single-precision compilable ICON. This MR:
* Introduces configure flag `--enable-single-precision` which sets working precision to single precision (sp).
* Adds BB builders for single-precision `levante_gcc_sp, levante_intel_hybrid_sp, levante_nag_sp, levante_gpu_nvhpc_sp, lumi_cpu_sp, mpimac_gcc_sp` in merge2rc (build-only).
* Misc changes for single-precision in 'mo_communication, mo_test_communication, mo_mpi, mo_bc_anthro_emission'.

Approved-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Merged-by: Ralf Mueller <ralf.mueller@dkrz.de>

+++ src/testbed/mo_test_communication.f90
@@ -1067,7 +1067,7 @@ CONTAINS
     REAL(sp), ALLOCATABLE, TARGET :: arr_sp_4d(:,:,:,:), ref_arr_sp_4d(:,:,:,:) !
     TYPE(t_ptr_3d_sp), ALLOCATABLE :: ptr_3d_arr_sp(:), ref_ptr_3d_arr_sp(:)
 #ifdef __PGI
-    REAL(dp), POINTER :: tmp_ptr_4d(:,:,:,:)
+    REAL(wp), POINTER :: tmp_ptr_4d(:,:,:,:)
 #endif
 
     CHARACTER(*), PARAMETER :: method_name = "mo_test_communication:sync_patch_array_testbed"
